pyproject.toml
[project]
name = "mtcat-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firebase-admin>=6.6.0",
    "google-api-python-client>=2.156.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",    
    "pandas>=2.2.3",
    "ruff>=0.8.4",
]
